Added generic IRestClient interface + JSON client
The client implementation is located in Plugins for the reason that
HttpClient on WP8 requires Microsoft.Http async library and I did not
want to introduce that dependency to all clients. I will try to put in
a ProtoBuf client at some point.

This client implementation depends on the serializer interfaces in the
core project.