Added Extended Master Detail Page
On iOS this will let you choose whether the menu is visible on iPad
when in landscape mode. Default is non-visible opposite to
MasterDetailPage on Forms.

Currently there is no extended functionality for Android or WP8 but the
renderers are there so if we ever add more functionality to the page
then we have them ready.