Fix bug where email sheet was not presented if top controller was modal
The window root view controller will not always be the one that iOS
considers the topmost. If a view controller is presented modally on top
of the root view, this modal should be used for presenting the email
sheet.

The code travels upwards through “PresentedViewController” until the
topmost is reached.